+++ backend/great-listens/src/main/java/com/database/greatlistens/controller/RatingsController.java
@@ -9,6 +9,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
 
+import com.database.greatlistens.Token.TokenManager;
 import com.database.greatlistens.model.Ratings;
 import com.database.greatlistens.service.RatingsService;
 
@@ -17,6 +18,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
 
 
 @Controller
@@ -27,34 +29,46 @@ public class RatingsController {
     private RatingsService ratingsService;
 
     @PostMapping("/create")
-    public ResponseEntity<String> createRating(@RequestBody Map<String, Object> requestBody) {
-        int bookId = (int) requestBody.get("book_id");
+    public ResponseEntity<String> createRating(@RequestHeader("Authorization") String token, @RequestBody Map<String, Object> requestBody) {
         String memId = (String) requestBody.get("mem_id");
+        if (!TokenManager.validateToken(token, memId)) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Invalid token");
+        }
+        int bookId = (int) requestBody.get("book_id");
         int rate = (int) requestBody.get("rate");
         ratingsService.createRating(bookId, memId, rate);
         return new ResponseEntity<>("Rating created successfully", HttpStatus.CREATED);
     }
 
     @GetMapping("/member")
-    public ResponseEntity<List<Ratings>> getAllRatingsMadeByMember(@RequestBody Map<String, Object> requestBody) {
+    public ResponseEntity getAllRatingsMadeByMember(@RequestHeader("Authorization") String token, @RequestBody Map<String, Object> requestBody) {
         String memId = (String) requestBody.get("mem_id");
+        if (!TokenManager.validateToken(token, memId)) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Invalid token");
+        }
         List<Ratings> ratings = ratingsService.getAllRatingsMadeByMember(memId);
         return new ResponseEntity<>(ratings, HttpStatus.OK);
     }
 
     @PutMapping("/update")
-    public ResponseEntity<String> updateRatings(@RequestBody Map<String, Object> requestBody) {
-        int bookId = (int) requestBody.get("book_id");
+    public ResponseEntity<String> updateRatings(@RequestHeader("Authorization") String token, @RequestBody Map<String, Object> requestBody) {
         String memId = (String) requestBody.get("mem_id");
+        if (!TokenManager.validateToken(token, memId)) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Invalid token");
+        }
+        int bookId = (int) requestBody.get("book_id");
         int rate = (int) requestBody.get("rate");
         ratingsService.updateRatings(bookId, memId, rate);
         return new ResponseEntity<>("Rating updated successfully", HttpStatus.OK);
     }
 
     @DeleteMapping("/delete")
-    public ResponseEntity<String> deleteRating(@RequestBody Map<String, Object> requestBody) {
-        int bookId = (int) requestBody.get("book_id");
+    public ResponseEntity<String> deleteRating(@RequestHeader("Authorization") String token, @RequestBody Map<String, Object> requestBody) {
         String memId = (String) requestBody.get("mem_id");
+        if (!TokenManager.validateToken(token, memId)) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Invalid token");
+        }
+        int bookId = (int) requestBody.get("book_id");
         ratingsService.deleteRating(bookId, memId);
         return new ResponseEntity<>("Rating deleted successfully", HttpStatus.OK);
     }
